code/requirements.txt
requests~=2.32.3
opencv-python~=4.11.0.86
pytesseract~=0.3.13
numpy~=2.2.2
pillow~=11.1.0
easyocr~=1.7.2
ollama~=0.4.7
textstat~=0.7.4
ollama-ocr
mpmath~=1.3.0
CairoSVG~=2.7.1
python-dotenv~=1.0.1
ocrmac~=1.0.0
pyobjc-framework-Vision~=11.0
pyobjc-framework-Cocoa~=11.0